Fix a topology bug in flavors support
This patch fixes an oversight in the addition of flavors support in the
amphora driver[1]. The amphora-agent configuration file was still getting
the topology selected in the configuration file as opposed to the
topology selected in the flavor.

This is an additional patch at the end of the chain as it leverages
changes that were made in later flavor patches that pass the flavor
into the taskflow flow.

A follow on patch will address spares pool amphora.

[1] https://review.openstack.org/#/c/621323

jira: http://172.18.211.163:8080/browse/NEUT-147
origin: https://review.openstack.org/#/c/632551/2
Change-Id: Iba62df76a8dc03b2c4a5bed18f9cf94f50f19e9d